+++ common/predictive-text/unit_tests/in_browser/CI.conf.js
@@ -87,12 +87,6 @@ module.exports = function(config) {
       browser: 'chrome',
       browser_version: '70.0'
     },
-    bs_ie_win: {
-      os: 'Windows',
-      os_version: '10',
-      browser: 'ie',
-      browser_version: '11.0'
-    },
     // On recent versions of Edge, launcher fails to start and/or stop Edge successfully
     //bs_edge_win: {
     //  os: 'Windows',